requirements.txt
matplotlib==3.8.4
numpy==1.26.4
openpyxl==3.1.5
Pillow==10.3.0
pytorch_crf==0.7.2
scikit_learn==1.5.2
torch==2.3.1
torchvision==0.18.1
transformers==4.30.2
sentencepiece==0.1.99
protobuf==3.20.3
